moved fixture definitions for data_mapping tests to tests/data_mapping/conftest.py
