+++ fast_carpenter/data_mapping/__init__.py
@@ -116,7 +116,6 @@ class DataMapping(abc.MutableMapping):
     def __contains__(self, name):
         in_extra = name in self._extra_variables
         in_connector = name in self._connector
-        print(f"{in_connector=}, {in_extra=}")
         return in_extra or in_connector
 
     def __getitem__(self, key):
@@ -145,6 +144,35 @@ class DataMapping(abc.MutableMapping):
     @property
     def num_entries(self):
         return self._connector.num_entries
+
+
+def __create_mapping_with_tree_connector__(input_file: FileLike, treenames: str, methods: str):
+    if len(treenames) > 1:
+        raise NotImplementedError(
+            "Multiple trees not supported by the TreeConnector. Please use a different connector.")
+    return DataMapping(
+        connector=DATA_CONNECTORS["tree"](
+            tree=input_file[treenames[0]], methods=ARRAY_METHODS[methods]),
+        methods=ARRAY_METHODS[methods],
+        indices=[IndexWithAliases({}), TokenMapIndex()],
+    )
+
+
+def __create_mapping_with_file_connector__(input_file: FileLike, treenames: str, methods: str):
+    return DataMapping(
+        connector=DATA_CONNECTORS["file"](file_handle=input_file, methods=ARRAY_METHODS[methods]),
+        methods=ARRAY_METHODS[methods],
+        indices=[IndexWithAliases({}), TokenMapIndex(), MultiTreeIndex(treenames)],
+    )
+
+
+def create_mapping(input_file: FileLike, treenames: str, methods: str, connector: str):
+    if connector == "tree":
+        return __create_mapping_with_tree_connector__(input_file, treenames, methods)
+    if connector == "file":
+        return __create_mapping_with_file_connector__(input_file, treenames, methods)
+
+    raise ValueError(f"Connector {connector} not supported by this function.")
 
 
 def create_ranged_mapping():
